+++ src/components/About/About.js
@@ -1,0 +1,13 @@
+import React from 'react';
+import styles from './About.module.css';
+
+const About = () => {
+  
+
+    return (
+      <div className = { styles.wrap }>
+          <h1 className = { styles.title }>About</h1>
+      </div>);
+}
+
+export default About;
