+++ src/components/App/App.module.css
@@ -1,11 +1,8 @@
 .wrap {
-    margin: auto;
-    width: 550px;
+    display: flex;
 }
 
-.title {
-    text-align: center;
-    font-family: Impact, Haettenschweiler, 'Arial Narrow Bold', sans-serif;
-    font-size: 60px;
-    color: darkcyan;
+.link {
+    text-decoration: none;
+    color:black;
 }
