+++ src/components/Contacts/Contacts.js
@@ -1,0 +1,14 @@
+import React from 'react';
+
+import styles from './Contacts.module.css';
+
+const Contacts = () => {
+  
+
+    return (
+      <div className = { styles.wrap }>
+        <h1 className = { styles.title }>Contacts</h1>
+      </div>);
+}
+
+export default Contacts;
