+++ src/components/Todo/Todo.module.css
@@ -1,0 +1,11 @@
+.wrap {
+    margin: auto;
+    width: 550px;
+}
+
+.title {
+    text-align: center;
+    font-family: Impact, Haettenschweiler, 'Arial Narrow Bold', sans-serif;
+    font-size: 60px;
+    color: darkcyan;
+}
